refactor: extract chat turn rows from AgentWorkspace
